+++ src/main/scala/tscfg/model.scala
@@ -43,7 +43,9 @@ object model {
 
   }
 
-  sealed abstract class Type
+  sealed abstract class Type {
+    def isObject: Boolean = this.isInstanceOf[ObjectAbsType]
+  }
 
   sealed abstract class BasicType extends Type
 
@@ -87,6 +89,7 @@ object model {
       optional: Boolean = false,
       default: Option[String] = None,
       defineCase: Option[DefineCase] = None,
+      docComments: List[String] = Nil,
       comments: Option[String] = None,
       parentClassMembers: Option[Map[String, model.AnnType]] = None,
   ) {